[NFC][Outlining] Consolidate code
Move the contents of stringify-walker-impl.h into stringify-walker.h,
and move everything not related to the base StringifyWalker class out of
stringify-walker.h and hash-stringify-walker.cpp and into Outlining.cpp.
Remove a few unit tests for functionality that is no longer exposed
outside of Outlining.cpp. These unit tests will eventually be replaced
with lit tests for the outlining pass.
