+++ src/passes/stringify-walker.h
@@ -17,15 +17,11 @@
 #ifndef wasm_passes_stringify_walker_h
 #define wasm_passes_stringify_walker_h
 
+#include <queue>
+
 #include "ir/iteration.h"
 #include "ir/module-utils.h"
-#include "ir/stack-utils.h"
-#include "ir/utils.h"
-#include "support/intervals.h"
-#include "support/suffix_tree.h"
-#include "wasm-ir-builder.h"
 #include "wasm-traversal.h"
-#include <queue>
 
 namespace wasm {
 
@@ -223,81 +219,123 @@ struct StringifyWalker
   void dequeueControlFlow();
 };
 
-} // namespace wasm
+// Implementation follows
 
-#include "stringify-walker-impl.h"
+// This walker supplies its own doWalkModule because it does not make sense to
+// walk anything besides defined functions.
+template<typename SubType>
+inline void StringifyWalker<SubType>::doWalkModule(Module* module) {
+  ModuleUtils::iterDefinedFunctions(
+    *module, [&](Function* func) { this->walkFunction(func); });
+}
 
-namespace wasm {
+template<typename SubType>
+inline void StringifyWalker<SubType>::doWalkFunction(Function* func) {
+  addUniqueSymbol(SeparatorReason::makeFuncStart(func));
+  Super::walk(func->body);
+  addUniqueSymbol(SeparatorReason::makeEnd());
+  flushControlFlowQueue();
+}
 
-// This custom hasher conforms to std::hash<Key>. Its purpose is to provide
-// a custom hash for if expressions, so the if-condition of the if expression is
-// not included in the hash for the if expression. This is needed because in the
-// binary format, the if-condition comes before and is consumed by the if. To
-// match the binary format, we hash the if condition before and separately from
-// the rest of the if expression.
-struct StringifyHasher {
-  size_t operator()(Expression* curr) const;
-};
+template<typename SubType>
+inline void StringifyWalker<SubType>::scan(SubType* self, Expression** currp) {
+  Expression* curr = *currp;
+  if (Properties::isControlFlowStructure(curr)) {
+    self->controlFlowQueue.push(curr);
+    self->pushTask(doVisitExpression, currp);
+    // The if-condition is a value child consumed by the if control flow, which
+    // makes the if-condition a true sibling rather than part of its contents in
+    // the binary format
+    for (auto*& child : ValueChildIterator(curr)) {
+      Super::scan(self, &child);
+    }
+  } else {
+    Super::scan(self, currp);
+  }
+}
 
-// This custom equator conforms to std::equal_to<Key>. Similar to
-// StringifyHasher, it's purpose is to not include the if-condition when
-// evaluating the equality of two if expressions.
-struct StringifyEquator {
-  bool operator()(Expression* lhs, Expression* rhs) const;
-};
+// This dequeueControlFlow is responsible for visiting the children expressions
+// of control flow.
+template<typename SubType> void StringifyWalker<SubType>::dequeueControlFlow() {
+  auto& queue = controlFlowQueue;
+  Expression* curr = queue.front();
+  queue.pop();
 
-struct HashStringifyWalker : public StringifyWalker<HashStringifyWalker> {
-  // After calling walkModule, this vector contains the result of encoding a
-  // wasm module as a string of uint32_t values. Each value represents either an
-  // Expression or a separator to mark the end of control flow.
-  std::vector<uint32_t> hashString;
-  // A monotonic counter used to ensure that unique expressions in the
-  // module are assigned a unique value in the hashString.
-  uint32_t nextVal = 0;
-  // A monotonic counter used to ensure that each separator in the
-  // module is assigned a unique value in the hashString.
-  int32_t nextSeparatorVal = -1;
-  // Contains a mapping of expression pointer to value to ensure we
-  // use the same value for matching expressions. A custom hasher and
-  // equator is provided in order to separate out evaluation of the if-condition
-  // when evaluating if expressions.
-  std::unordered_map<Expression*, uint32_t, StringifyHasher, StringifyEquator>
-    exprToCounter;
-  std::vector<Expression*> exprs;
+  // TODO: Issue #5796, Make a ControlChildIterator
+  switch (curr->_id) {
+    case Expression::Id::BlockId: {
+      auto* block = curr->cast<Block>();
+      addUniqueSymbol(SeparatorReason::makeBlockStart(block));
+      for (auto& child : block->list) {
+        Super::walk(child);
+      }
+      addUniqueSymbol(SeparatorReason::makeEnd());
+      break;
+    }
+    case Expression::Id::IfId: {
+      auto* iff = curr->cast<If>();
+      addUniqueSymbol(SeparatorReason::makeIfStart(iff));
+      Super::walk(iff->ifTrue);
+      if (iff->ifFalse) {
+        addUniqueSymbol(SeparatorReason::makeElseStart());
+        Super::walk(iff->ifFalse);
+      }
+      addUniqueSymbol(SeparatorReason::makeEnd());
+      break;
+    }
+    case Expression::Id::TryId: {
+      auto* tryy = curr->cast<Try>();
 
-  void addUniqueSymbol(SeparatorReason reason);
-  void visitExpression(Expression* curr);
-  // Converts the idx from relative to the beginning of the program to
-  // relative to its enclosing function, and returns the name of its function.
-  std::pair<uint32_t, Name> makeRelative(uint32_t idx) const;
+      addUniqueSymbol(SeparatorReason::makeTryStart(tryy));
+      Super::walk(tryy->body);
+      for (size_t i = 0; i < tryy->catchBodies.size(); i++) {
+        if (tryy->hasCatchAll() && i == tryy->catchBodies.size() - 1) {
+          addUniqueSymbol(SeparatorReason::makeCatchAllStart());
+        } else {
+          addUniqueSymbol(SeparatorReason::makeCatchStart(tryy->catchTags[i]));
+        }
+        Super::walk(tryy->catchBodies[i]);
+      }
+      addUniqueSymbol(SeparatorReason::makeEnd());
+      break;
+    }
+    case Expression::Id::LoopId: {
+      auto* loop = curr->cast<Loop>();
+      addUniqueSymbol(SeparatorReason::makeLoopStart(loop));
+      Super::walk(loop->body);
+      addUniqueSymbol(SeparatorReason::makeEnd());
+      break;
+    }
+    case Expression::Id::TryTableId: {
+      auto* tryt = curr->cast<TryTable>();
+      addUniqueSymbol(SeparatorReason::makeTryTableStart(tryt));
+      Super::walk(tryt->body);
+      addUniqueSymbol(SeparatorReason::makeEnd());
+      break;
+    }
+    default: {
+      assert(Properties::isControlFlowStructure(curr));
+      WASM_UNREACHABLE("unexpected expression");
+    }
+  }
+}
 
-private:
-  // Contains the indices that mark the start of each function.
-  std::set<uint32_t> funcIndices;
-  // Maps the start idx of each function to the function name.
-  std::map<uint32_t, Name> idxToFuncName;
-};
+template<typename SubType>
+void StringifyWalker<SubType>::doVisitExpression(SubType* self,
+                                                 Expression** currp) {
+  Expression* curr = *currp;
+  self->visit(curr);
+}
 
-using Substrings = std::vector<SuffixTree::RepeatedSubstring>;
-
-// Functions that filter vectors of SuffixTree::RepeatedSubstring
-struct StringifyProcessor {
-  static Substrings repeatSubstrings(std::vector<uint32_t>& hashString);
-  static Substrings dedupe(const Substrings& substrings);
-  static Substrings filterOverlaps(const Substrings& substrings);
-  // Filter is the general purpose function backing subsequent filter functions.
-  // It can be used directly, but generally prefer a wrapper function
-  // to encapsulate your condition and make it available for tests.
-  static Substrings filter(const Substrings& substrings,
-                           const std::vector<Expression*>& exprs,
-                           std::function<bool(const Expression*)> condition);
-  static Substrings filterLocalSets(const Substrings& substrings,
-                                    const std::vector<Expression*>& exprs);
-  static Substrings filterLocalGets(const Substrings& substrings,
-                                    const std::vector<Expression*>& exprs);
-  static Substrings filterBranches(const Substrings& substrings,
-                                   const std::vector<Expression*>& exprs);
-};
+template<typename SubType>
+inline void StringifyWalker<SubType>::addUniqueSymbol(SeparatorReason reason) {
+  // TODO: Add the following static_assert when the compilers running our GitHub
+  // actions are updated enough to know that this is a constant condition:
+  // static_assert(&StringifyWalker<SubType>::addUniqueSymbol !=
+  // &SubType::addUniqueSymbol);
+  auto self = static_cast<SubType*>(this);
+  self->addUniqueSymbol(reason);
+}
 
 } // namespace wasm
